requirements.txt
absl-py==0.2.0
aiohttp==1.0.5
astor==0.6.2
async-timeout==2.0.1
bleach==1.5.0
chardet==3.0.4
discord.py==0.16.12
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.5
Markdown==2.6.11
multidict==4.2.0
numpy==1.14.2
protobuf==3.5.2.post1
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
tensorboard==1.7.0
tensorflow-gpu==1.7.0
termcolor==1.1.0
textgenrnn==1.0
websockets==3.4
Werkzeug==0.14.1
